refactor(cli): lift Prometheus reader and metrics loader into metrics_source
Move the per-rule `*_matches_by_rule_total` text-exposition parser and the
file-or-URL metrics loader out of `rule scorecard` into a shared
`crate::metrics_source` module so other rule commands can join production
fire volume without reimplementing the format. Behavior-neutral: scorecard
re-exports the loader under its existing `inputs::` names and maps the
shared error into its own exit-code intent; the fuzz target now includes
the parser from its new path.
