+++ crates/rsigma-cli/src/main.rs
@@ -5,6 +5,7 @@ mod daemon;
 pub(crate) mod exit_code;
 mod fix;
 pub(crate) mod logsource_opts;
+pub(crate) mod metrics_source;
 pub(crate) mod output;
 pub(crate) mod rule_meta;
